requirements.txt
fastapi
uvicorn[standard]
pandas
PyYAML
langchain
langchain-openai
python-dotenv